update totalbilling hours

+++ final-project/TimeSheetServer/src/main/java/com/example/timesheetserver/domain/SummaryDomain.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @Getter
 public class SummaryDomain {
     private String weekEnding;
-    private int totalHours;
+    private double totalHours;
     private String submissionStatus;
     private String approvalStatus;
     private String option;
